Separate memory and kmalloc modules
memory_raw module contains memory_init and page-wide operations, while
kmalloc is the kernel memory allocator. memory_raw will be used for
other modules, so it makes sense to separate it out. memory.h becomes an
umbrella header that includes all memory-related headers.

Tested by compiling and running, and everything seems to work just as
before.
